refactor: Getting a string from an error is a util function
