+++ client/src/pages/UserRegistration.tsx
@@ -4,6 +4,7 @@ import './UserRegistration.css';
 import { UserRegisterDTO, UserService } from '../api/user.api';
 import { AxiosError } from 'axios';
 import { useNavigate } from 'react-router-dom';
+import { get_validation_error_readable } from '../util/http';
 
 export const UserRegistration = () => {
     let navigate = useNavigate();
@@ -43,20 +44,7 @@ export const UserRegistration = () => {
         UserService.RegisterRegularUser(dto).then(() => {
             navigate("/login");
         }).catch((err: AxiosError) => {
-            try {
-                let err_msg = (err.response?.data as any)["detail"]["message"];
-                if (Array.isArray(err_msg)) {
-                    setError(err_msg[0]["msg"]);
-                } else if (typeof err_msg == "string") {
-                    setError(err_msg);
-                } else {
-                    setError((err.response?.data as any)["detail"]["message"]);
-                }
-            } catch (e) {
-                setError("Invalid user input");
-                console.error(e);
-                console.error(err.response?.data as any);
-            }
+            setError(get_validation_error_readable(err));
         });
     };
